refactor : DAO 역할을 하는 Mapper 클래스들을 comparingprices 패키지로 이동 - 중복되는 ApartmentMapper 제거 - ApartmentPriceMapper는 comparingprices 패키지에서도 사용할 수 있으므로 이동 => 중복될 수 있는 로직들은 메인 패키지인 comparingprices 패키지로 이동하여 utils만 comparingprices 패키지에 의존하도록 한다.
