+++ src/main/java/kancho/realestate/utils/api/StorePricesMain.java
@@ -4,7 +4,7 @@ import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = {"kancho.realestate"})
 public class StorePricesMain {
 
 	public static void main(String[] args){
