+++ src/main/java/kancho/realestate/utils/excel/StoreApartmentMain.java
@@ -5,7 +5,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 
 @SpringBootApplication(scanBasePackages = {"kancho.realestate"})
-
 public class StoreApartmentMain {
 
 	public static void main(String[] args){
