+++ src/main/resources/mybatis-mapper/ApartmentPriceMapper.xml
@@ -3,7 +3,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="kancho.realestate.utils.api.storeaprtment.mapper.ApartmentPriceMapper">
+<mapper namespace="kancho.realestate.comparingprices.repository.ApartmentPriceMapper">
 
     <insert id="save" parameterType="ApartmentPrice">
         insert
