--- src/main/resources/mybatis-mapper/utils-api-storeaprtment-mapper/ApartmentMapper.xml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="kancho.realestate.utils.api.storeaprtment.mapper.ApartmentMapper">
-
-    <select id="findAll" resultType="Apartment">
-        select
-        id
-        , regional_code
-        , city
-        , gu
-        , dong
-        , jibun
-        , bonbun
-        , bubun
-        , apartment_name
-        , build_year
-        , road_address
-        from apartment
-        where delete_dttm is null
-    </select>
-
-    <insert id="save" parameterType="Apartment">
-        insert
-        into apartment
-        ( regional_code
-        , city
-        , gu
-        , dong
-        , jibun
-        , bonbun
-        , bubun
-        , apartment_name
-        , build_year
-        , road_address)
-        values ( #{regionalCode}
-        , #{city}
-        , #{gu}
-        , #{dong}
-        , #{jibun}
-        , #{bonbun}
-        , #{bubun}
-        , #{apartmentName}
-        , #{buildYear}
-        , #{roadAddress}
-        )
-    </insert>
-
-</mapper>
